Documents/RITIKAMTP/TCSS-CRISIS-DATASET/requirements.txt
nltk
scikit-learn
numpy
spacy
transformers
torch
pandas
pillow
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
